chore(frame): move frame stack modifiers in a separate frame-stack module
